Adding alert dialog for delete action

+++ src/app/(main)/posts/columns.tsx
@@ -8,12 +8,23 @@ import {
   DropdownMenu,
   DropdownMenuContent,
   DropdownMenuItem,
-  DropdownMenuLabel,
+  DropdownMenuSeparator,
   DropdownMenuTrigger,
 } from "~/components/ui/dropdown-menu";
 import { api } from "~/trpc/react";
 import { type SelectPost } from "~/types/posts";
 import { useSession } from "next-auth/react";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+} from "~/components/ui/alert-dialog";
+import { useState } from "react";
 
 export const columns: ColumnDef<SelectPost>[] = [
   {
@@ -58,33 +69,51 @@ function PostsActionsCell({ row }: CellContext<SelectPost, unknown>) {
   const { mutate: deletePost, isPending } = api.post.deletePost.useMutation();
 
   const { data: session } = useSession();
+  const [openDeleteDialog, setOpenDeleteDialog] = useState(false);
 
   const isAuthenticated = Boolean(session?.user?.id);
 
   return (
-    <DropdownMenu>
-      <DropdownMenuTrigger asChild>
-        <Button variant="ghost" className="h-8 w-8 p-0">
-          <span className="sr-only">Open menu</span>
-          <MoreHorizontal className="h-4 w-4" />
-        </Button>
-      </DropdownMenuTrigger>
-      <DropdownMenuContent align="end">
-        <DropdownMenuLabel>Actions</DropdownMenuLabel>
-        <DropdownMenuItem
-          onClick={() => navigator.clipboard.writeText(post.id.toString())}
-        >
-          Copy Post ID
-        </DropdownMenuItem>
-        {isAuthenticated && (
+    <AlertDialog open={openDeleteDialog} onOpenChange={setOpenDeleteDialog}>
+      <DropdownMenu modal={false}>
+        <DropdownMenuTrigger asChild>
+          <Button variant="ghost" className="h-8 w-8 p-0">
+            <span className="sr-only">Open menu</span>
+            <MoreHorizontal className="h-4 w-4" />
+          </Button>
+        </DropdownMenuTrigger>
+        <DropdownMenuContent align="end">
           <DropdownMenuItem
+            onClick={() => navigator.clipboard.writeText(post.id.toString())}
+          >
+            Copy Post ID
+          </DropdownMenuItem>
+          <DropdownMenuSeparator />
+          {isAuthenticated && (
+            <DropdownMenuItem onClick={() => setOpenDeleteDialog(true)}>
+              Delete Post
+            </DropdownMenuItem>
+          )}
+        </DropdownMenuContent>
+      </DropdownMenu>
+      <AlertDialogContent>
+        <AlertDialogHeader>
+          <AlertDialogTitle>Are you absolutely sure?</AlertDialogTitle>
+          <AlertDialogDescription>
+            This action cannot be undone. This will permanently delete your
+            account and remove your data from our servers.
+          </AlertDialogDescription>
+        </AlertDialogHeader>
+        <AlertDialogFooter>
+          <AlertDialogCancel>Cancel</AlertDialogCancel>
+          <AlertDialogAction
             onClick={() => deletePost(post.id)}
             disabled={isPending}
           >
-            Delete Post
-          </DropdownMenuItem>
-        )}
-      </DropdownMenuContent>
-    </DropdownMenu>
+            Continue
+          </AlertDialogAction>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
   );
 }
